Update to use Reference Class OOP system
Instead of keeping data, registry, counter, etc, separate, they are now combined into a class using the Reference class system. The reference class system was chosen as opposed to an S3 or S4 class because it seems to be appropriate for objects that update themselves (they don't copy-on-modify), which hopefully will lead to performance improvements.

This change required major refactoring of most all functions, but will greatly simplify the workflow.